Convert Java classloading code into a mixin
