+++ lib/msf/core/exploit/java.rb
@@ -16,6 +16,7 @@
 require 'msf/core'
 require 'msf/core/exploit/java/rmi/util'
 require 'msf/core/exploit/java/rmi/client'
+require 'msf/core/exploit/java/classloading'
 
 module Msf
 module Exploit::Java
